Convert Assert<String> extensions into Assert<CharSequence> extensions.
