+++ assertk/src/commonMain/kotlin/assertk/assertions/string.kt
@@ -6,15 +6,6 @@ import assertk.assertions.support.fail
 import assertk.assertions.support.show
 
 /**
- * Asserts the string has the expected number of lines.
- */
-fun Assert<String>.hasLineCount(lineCount: Int) = given { actual ->
-    val actualLineCount = actual.lines().size
-    if (actualLineCount == lineCount) return
-    expected("to have line count:${show(lineCount)} but was:${show(actualLineCount)}")
-}
-
-/**
  * Asserts the string is equal to the expected string.
  * @param ignoreCase true to compare ignoring case, the default if false.
  * @see [isNotEqualTo]
@@ -36,86 +27,4 @@ fun Assert<String?>.isNotEqualTo(other: String?, ignoreCase: Boolean = false) = 
     } else {
         expected("to not be equal to:${show(actual)}")
     }
-}
-
-/**
- * Asserts the string contains the expected substring.
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.contains(expected: CharSequence, ignoreCase: Boolean = false) = given { actual ->
-    if (actual.contains(expected, ignoreCase)) return
-    expected("to contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string contains the expected substring(s).
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.contains(vararg expected: CharSequence, ignoreCase: Boolean = false) = given { actual ->
-    if (expected.all { actual.contains(it, ignoreCase) }) return
-    expected("to contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string contains the expected strings.
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.contains(expected: Iterable<CharSequence>, ignoreCase: Boolean = false) = given { actual ->
-    if (expected.all { actual.contains(it, ignoreCase) }) return
-    expected("to contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string does not contain the specified string.
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.doesNotContain(expected: CharSequence, ignoreCase: Boolean = false) = given { actual ->
-    if (!actual.contains(expected, ignoreCase)) return
-    expected("to not contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string does not contain the specified string(s).
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.doesNotContain(vararg expected: CharSequence, ignoreCase: Boolean = false) = given { actual ->
-    if (expected.none { actual.contains(it, ignoreCase) }) return
-    expected("to not contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string does not contain the specified strings.
- * @param ignoreCase true to compare ignoring case, the default if false.
- */
-fun Assert<String>.doesNotContain(expected: Iterable<CharSequence>, ignoreCase: Boolean = false) = given { actual ->
-    if (expected.none { actual.contains(it, ignoreCase) }) return
-    expected("to not contain:${show(expected)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string starts with the expected string.
- * @param ignoreCase true to compare ignoring case, the default if false.
- * @see [endsWith]
- */
-fun Assert<String>.startsWith(other: String, ignoreCase: Boolean = false) = given { actual ->
-    if (actual.startsWith(other, ignoreCase)) return
-    expected("to start with:${show(other)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string ends with the expected string.
- * @param ignoreCase true to compare ignoring case, the default if false.
- * @see [startsWith]
- */
-fun Assert<String>.endsWith(other: String, ignoreCase: Boolean = false) = given { actual ->
-    if (actual.endsWith(other, ignoreCase)) return
-    expected("to end with:${show(other)} but was:${show(actual)}")
-}
-
-/**
- * Asserts the string matches the expected regular expression.
- */
-fun Assert<String>.matches(regex: Regex) = given { actual ->
-    if (actual.matches(regex)) return
-    expected("to match:${show(regex)} but was:${show(actual)}")
 }
